+++ tests/test_snappy_itransfer_variant_calling.py
@@ -1,4 +1,4 @@
-"""Tests for ``cubi_sak.snappy.itransfer_ngs_mapping``.
+"""Tests for ``cubi_sak.snappy.itransfer_variant_calling``.
 
 We only run some smoke tests here.
 """
@@ -13,10 +13,10 @@ from pyfakefs import fake_filesystem
 from cubi_sak.__main__ import setup_argparse, main
 
 
-def test_run_snappy_itransfer_ngs_mapping_help(capsys):
+def test_run_snappy_itransfer_variant_calling_help(capsys):
     parser, subparsers = setup_argparse()
     with pytest.raises(SystemExit) as e:
-        parser.parse_args(["snappy", "itransfer-ngs-mapping", "--help"])
+        parser.parse_args(["snappy", "itransfer-variant-calling", "--help"])
 
     assert e.value.code == 0
 
@@ -25,11 +25,11 @@ def test_run_snappy_itransfer_ngs_mapping_help(capsys):
     assert not res.err
 
 
-def test_run_snappy_itransfer_ngs_mapping_nothing(capsys):
+def test_run_snappy_itransfer_variant_calling_nothing(capsys):
     parser, subparsers = setup_argparse()
 
     with pytest.raises(SystemExit) as e:
-        parser.parse_args(["snappy", "itransfer-ngs-mapping"])
+        parser.parse_args(["snappy", "itransfer-variant-calling"])
 
     assert e.value.code == 2
 
@@ -38,14 +38,14 @@ def test_run_snappy_itransfer_ngs_mapping_nothing(capsys):
     assert res.err
 
 
-def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
+def test_run_snappy_itransfer_variant_calling_smoke_test(mocker):
     fake_base_path = "/base/path"
     dest_path = "/irods/dest"
     tsv_path = os.path.join(os.path.dirname(__file__), "data", "germline.out")
     argv = [
         "--verbose",
         "snappy",
-        "itransfer-ngs-mapping",
+        "itransfer-variant-calling",
         "--base-path",
         fake_base_path,
         tsv_path,
@@ -57,18 +57,25 @@ def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
     fs = fake_filesystem.FakeFilesystem()
 
     fake_file_paths = []
-    for member in ("index", "father", "mother"):
+    for member in ("index",):
         for ext in ("", ".md5"):
             fake_file_paths.append(
-                "%s/ngs_mapping/output/bwa.%s-N1-DNA1-WES1/out/%s-N1-DNA1-WES1.bam%s"
+                "%s/variant_calling/output/bwa.gatk_hc.%s-N1-DNA1-WES1/out/bwa.gatk_hc.%s-N1-DNA1-WES1.vcf.gz%s"
                 % (fake_base_path, member, member, ext)
             )
             fs.create_file(fake_file_paths[-1])
             fake_file_paths.append(
-                "%s/ngs_mapping/output/bwa.%s-N1-DNA1-WES1/log/bwa.%s-N1-DNA1-WES1.log%s"
+                "%s/variant_calling/output/bwa.gatk_hc.%s-N1-DNA1-WES1/out/bwa.gatk_hc.%s-N1-DNA1-WES1.vcf.gz.tbi%s"
                 % (fake_base_path, member, member, ext)
             )
             fs.create_file(fake_file_paths[-1])
+            fake_file_paths.append(
+                "%s/variant_calling/output/bwa.gatk_hc.%s-N1-DNA1-WES1/log/bwa.gatk_hc.%s-N1-DNA1-WES1.log%s"
+                % (fake_base_path, member, member, ext)
+            )
+            fs.create_file(fake_file_paths[-1])
+
+    print("\n".join(fake_file_paths))
 
     # Remove index's log MD5 file again so it is recreated.
     fs.remove(fake_file_paths[3])
@@ -76,7 +83,7 @@ def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
     fake_os = fake_filesystem.FakeOsModule(fs)
     mocker.patch("glob.os", fake_os)
     mocker.patch("cubi_sak.snappy.itransfer_common.os", fake_os)
-    mocker.patch("cubi_sak.snappy.itransfer_ngs_mapping.os", fake_os)
+    mocker.patch("cubi_sak.snappy.itransfer_variant_calling.os", fake_os)
 
     mock_check_output = mock.mock_open()
     mocker.patch("cubi_sak.snappy.itransfer_common.check_output", mock_check_output)
@@ -101,7 +108,7 @@ def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
 
     assert mock_check_call.call_count == 1
     mock_check_call.assert_called_once_with(
-        ["md5sum", "bwa.index-N1-DNA1-WES1.log"],
+        ["md5sum", "bwa.gatk_hc.index-N1-DNA1-WES1.vcf.gz"],
         cwd=os.path.dirname(fake_file_paths[3]),
         stdout=ANY,
     )
@@ -109,10 +116,12 @@ def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
     assert mock_check_output.call_count == len(fake_file_paths) * 2
     for path in fake_file_paths:
         mapper_index, rel_path = os.path.relpath(
-            path, os.path.join(fake_base_path, "ngs_mapping/output")
+            path, os.path.join(fake_base_path, "variant_calling/output")
         ).split("/", 1)
         _mapper, index = mapper_index.rsplit(".", 1)
-        remote_path = os.path.join(dest_path, index, "ngs_mapping", args.remote_dir_date, rel_path)
+        remote_path = os.path.join(
+            dest_path, index, "variant_calling", args.remote_dir_date, rel_path
+        )
         expected_mkdir_argv = ["imkdir", "-p", os.path.dirname(remote_path)]
         expected_irsync_argv = ["irsync", "-a", "-K", path, "i:%s" % remote_path]
         mock_check_output.assert_any_call(expected_mkdir_argv)